requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-dynamic-fixture==3.1.2
django-extensions==3.1.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
